refactor: migrate back to pure composite
the idea with visitors turned out to be not as good as expected.
It complicates things a lot. This is because it
requires different context objects for different levels of tree.
This makes the combination of composite and visitor not worth it
as the work of tree traversal falls on visitor.

This reverts commit 3cf61174fc89cfc282c3f04d439f520da76c828a.